Caching ArrayFire and Cmake

+++ .CI/travis/install-arrayfire_osx.sh
@@ -5,11 +5,12 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#if [ ! -d "/opt/arrayfire" ]; then
+if [ ! -d "${TRAVIS_BUILD_DIR/arrayfire}" ]; then
+    mkdir -p arrayfire && cd arrayfire
+    wget http://arrayfire.s3.amazonaws.com/3.5.1/ArrayFire-no-gl-v3.5.1_OSX.pkg
+    sudo mkdir /opt/arrayfire
 
-wget http://arrayfire.s3.amazonaws.com/3.5.1/ArrayFire-no-gl-v3.5.1_OSX.pkg
-sudo mkdir /opt/arrayfire
-
-sudo installer -pkg ArrayFire-no-gl-v3.5.1_OSX.pkg -target /
-rm ArrayFire-no-gl-v3.5.1_OSX.pkg
-#fi
+    sudo installer -pkg ArrayFire-no-gl-v3.5.1_OSX.pkg -target /
+    rm ArrayFire-no-gl-v3.5.1_OSX.pkg
+    cd ..
+fi
